Move ModListIcons class into model/mod_list_item.py
